fix: handle memcap_pressure metrics
A recent change in suricata 7 fixes logging of stats that may be
"unscoped"
see: https://redmine.openinfosecfoundation.org/issues/6094

Because of this, suricata_exporter was not able to process those stats.
This change is specific to `memcap_pressure` and `memcap_pressure_max`.

It allows for handling those stats when "unscoped" like uptime or when
found within flow manager thread context.
This also updates the unit tests to verify that those stats are being
discovered/counted.